feat: improve layouts
