+++ examples/video-room/screens/room/index.tsx
@@ -1,24 +1,59 @@
-import { StyleSheet } from 'react-native';
+import { FlatList, ListRenderItemInfo, StyleSheet, View } from 'react-native';
 import { RootScreenProps } from '../../navigation/RootNavigation';
-import { SafeAreaView } from 'react-native-safe-area-context';
-import { VideosGrid } from '../../components/VideosGrid';
+import { useConnection, usePeers } from '@fishjam-cloud/react-native-client';
+import { parsePeersToTracks } from '../../utils';
+import { useCallback, useEffect } from 'react';
+import { GridTrack } from '../../types';
+import { VideosGridItem } from '../../components/VideosGridItem';
 
 export type RoomScreenProps = RootScreenProps<'Room'>;
 
 const RoomScreen = () => {
+  const { leaveRoom } = useConnection();
+  const { localPeer, remotePeers } = usePeers();
+  const videoTracks = parsePeersToTracks(localPeer, remotePeers);
+
+  const keyExtractor = useCallback((item: GridTrack) => item.peerId, []);
+
+  const renderItem = useCallback(
+    ({ item }: ListRenderItemInfo<GridTrack>) => <VideosGridItem peer={item} />,
+    [],
+  );
+
+  useEffect(() => {
+    return () => {
+      leaveRoom();
+    };
+  }, [leaveRoom]);
+
   return (
-    <SafeAreaView style={styles.container}>
-      <VideosGrid />
-    </SafeAreaView>
+    <View style={styles.container}>
+      <FlatList
+        data={videoTracks}
+        keyExtractor={keyExtractor}
+        renderItem={renderItem}
+        numColumns={2}
+        contentContainerStyle={styles.contentContainerStyle}
+        columnWrapperStyle={styles.columnWrapperStyle}
+      />
+    </View>
   );
 };
+
+export default RoomScreen;
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
     paddingBottom: 32,
+  },
+  contentContainerStyle: {
+    justifyContent: 'center',
+    paddingHorizontal: 16,
+    paddingTop: 16,
+    gap: 16,
+  },
+  columnWrapperStyle: {
     gap: 16,
   },
 });
-
-export default RoomScreen;
